refactor: move prisma to its own package
