+++ .npmrc
@@ -1,0 +1,1 @@
+public-hoist-pattern[]=*prisma*
