+++ packages/prisma/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@lagon/prisma",
+  "version": "1.0.0",
+  "private": true,
+  "main": "index.ts",
+  "scripts": {
+    "postinstall": "prisma generate"
+  },
+  "dependencies": {
+    "@prisma/client": "^3.15.2"
+  },
+  "devDependencies": {
+    "prisma": "^3.15.2"
+  }
+}
