+++ packages/website/lib/trpc/deploymentsRouter.ts
@@ -1,5 +1,5 @@
 import { removeDeployment, setCurrentDeployment } from 'lib/api/deployments';
-import prisma from 'lib/prisma';
+import prisma from '@lagon/prisma';
 import * as trpc from '@trpc/server';
 import { createRouter } from 'pages/api/trpc/[trpc]';
 import { z } from 'zod';
