+++ packages/website/lib/trpc/organizationsRouter.ts
@@ -5,7 +5,7 @@ import {
   ORGANIZATION_NAME_MAX_LENGTH,
   ORGANIZATION_NAME_MIN_LENGTH,
 } from 'lib/constants';
-import prisma from 'lib/prisma';
+import prisma from '@lagon/prisma';
 import { createRouter } from 'pages/api/trpc/[trpc]';
 import { z } from 'zod';
