+++ packages/website/next.config.js
@@ -5,6 +5,7 @@
 
 // eslint-disable-next-line
 const { withSentryConfig } = require('@sentry/nextjs');
+const withTM = require('next-transpile-modules')(['@lagon/prisma']);
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
@@ -25,4 +26,4 @@ const sentryWebpackPluginOptions = {
 
 // Make sure adding Sentry options is the last code to run before exporting, to
 // ensure that your source maps include changes from all other Webpack plugins
-module.exports = withSentryConfig(nextConfig, sentryWebpackPluginOptions);
+module.exports = withTM(withSentryConfig(nextConfig, sentryWebpackPluginOptions));
