+++ packages/website/package.json
@@ -9,16 +9,15 @@
     "lint": "next lint",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "postinstall": "prisma generate",
     "clickhouse": "node scripts/clickhouse.mjs"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.100.0",
     "@headlessui/react": "^1.6.4",
     "@heroicons/react": "^1.0.6",
+    "@lagon/prisma": "workspace:1.0.0",
     "@monaco-editor/react": "^4.4.5",
     "@next-auth/prisma-adapter": "^1.0.3",
-    "@prisma/client": "^3.15.2",
     "@radix-ui/react-alert-dialog": "^0.1.7",
     "@radix-ui/react-tabs": "^0.1.5",
     "@scaleway/use-random-name": "^0.2.3",
@@ -35,6 +34,7 @@
     "monaco-editor": "^0.33.0",
     "next": "12.1.6",
     "next-auth": "^4.3.4",
+    "next-transpile-modules": "^9.0.0",
     "react": "18.1.0",
     "react-chartjs-2": "^4.2.0",
     "react-dom": "18.1.0",
@@ -63,7 +63,6 @@
     "eslint-config-next": "12.1.6",
     "eslint-plugin-storybook": "^0.5.12",
     "postcss": "^8.4.14",
-    "prisma": "^3.15.2",
     "tailwindcss": "^3.0.24",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
     "typescript": "^4.6.4"
