+++ packages/website/pages/api/auth/[...nextauth].ts
@@ -1,7 +1,7 @@
 import NextAuth, { Session } from 'next-auth';
 import GithubProvider from 'next-auth/providers/github';
 import { PrismaAdapter } from '@next-auth/prisma-adapter';
-import prisma from 'lib/prisma';
+import prisma from '@lagon/prisma';
 import apiHandler from 'lib/api';
 import * as Sentry from '@sentry/nextjs';
